server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
anthropic>=0.28.0
stripe>=9.0.0
PyJWT>=2.8.0
pymysql>=1.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.0.0
